requirements.txt
bs4  # BeautifulSoup
numpy
opencv-python-headless  # OpenCV
Pillow
tqdm
